Update code for function spatialJoin()
The "spatialJoin" function was giving me trouble. It looks like you had a slightly different version of the function in your sample product. I added that to my code and got it to work. I plugged that working code in here and kept the comments from what you had in the existing instruction code.